master/Projets/EG/backEG/ecogroup/requirements.txt
Django==3.1.1
Pillow==7.2.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
django-rest-passwordreset==1.1.0
django-filter==2.4.0
django-heroku==0.3.1
gunicorn==20.0.4
